chore: upgrading firebase BOM
