+++ build.gradle
@@ -7,7 +7,7 @@ buildscript {
         hiltVersion = '2.45'
         hiltWorkVersion = '1.0.0'
         workRuntimeVersion='2.7.1'
-        firebaseVersion = '28.4.2'
+        firebaseVersion = '32.1.1'
         roomVersion = '2.4.3'
         jetpackVersion = '1.8.0'
         appCompatVersion = '1.4.2'
