+++ wallet/src/de/schildbach/wallet/data/StaticFeeLoader.java
@@ -1,0 +1,101 @@
+/*
+ * Copyright the original author or authors.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package de.schildbach.wallet.data;
+
+import android.content.Context;
+import android.content.res.AssetManager;
+
+import androidx.loader.content.AsyncTaskLoader;
+
+import com.google.common.base.Charsets;
+
+import org.bitcoinj.core.Coin;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.util.HashMap;
+import java.util.Map;
+
+import de.schildbach.wallet.Constants;
+
+/**
+ * @author Andreas Schildbach
+ */
+public class StaticFeeLoader extends AsyncTaskLoader<Map<FeeCategory, Coin>> {
+    private final AssetManager assets;
+
+    private static final Logger log = LoggerFactory.getLogger(StaticFeeLoader.class);
+
+    public StaticFeeLoader(final Context context) {
+        super(context);
+        this.assets = context.getAssets();
+    }
+
+    @Override
+    protected void onStartLoading() {
+        super.onStartLoading();
+        forceLoad();
+    }
+
+    @Override
+    public Map<FeeCategory, Coin> loadInBackground() {
+        try {
+            return parseFees(assets.open(Constants.Files.FEES_FILENAME));
+        } catch (final IOException x) {
+            // Should not happen
+            throw new RuntimeException(x);
+        }
+    }
+
+    private static Map<FeeCategory, Coin> parseFees(final InputStream is) throws IOException {
+        final Map<FeeCategory, Coin> dynamicFees = new HashMap<FeeCategory, Coin>();
+        BufferedReader reader = null;
+        String line = null;
+        try {
+            reader = new BufferedReader(new InputStreamReader(is, Charsets.US_ASCII));
+            while (true) {
+                line = reader.readLine();
+                if (line == null)
+                    break;
+                line = line.trim();
+                if (line.length() == 0 || line.charAt(0) == '#')
+                    continue;
+
+                final String[] fields = line.split("=");
+                try {
+                    final FeeCategory category = FeeCategory.valueOf(fields[0]);
+                    final Coin rate = Coin.valueOf(Long.parseLong(fields[1]));
+                    dynamicFees.put(category, rate);
+                } catch (IllegalArgumentException x) {
+                    log.warn("Cannot parse line, ignoring: '" + line + "'", x);
+                }
+            }
+        } catch (final Exception x) {
+            throw new RuntimeException("Error while parsing: '" + line + "'", x);
+        } finally {
+            if (reader != null)
+                reader.close();
+            is.close();
+        }
+        return dynamicFees;
+    }
+}
